# openapi-crm-springboot.yaml
openapi: 3.0.3
info:
  title: CRM Backend API
  version: 1.0.0
servers:
  - url: /api/v1
paths:
...
  /teams:
    get:
      tags: [Teams]
      summary: List teams
      responses:
        '200':
          description: Teams
    post:
      tags: [Teams]
      summary: Create team
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                teamName:
                  type: string
      responses:
        '201':
          description: Team created
  /teams/{id}:
    get:
      tags: [Teams]
      summary: Get team by id
      parameters:
        - $ref: '#/components/parameters/IdParam'
      responses:
        '200':
          description: Team
    put:
      tags: [Teams]
      summary: Update team
      parameters:
        - $ref: '#/components/parameters/IdParam'
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                teamName:
                  type: string
      responses:
        '200':
          description: Team updated
    delete:
      tags: [Teams]
      summary: Delete team
      parameters:
        - $ref: '#/components/parameters/IdParam'
      responses:
        '200':
          description: Team deleted
  /team-members:
    get:
      tags: [Team Members]
      summary: List team members
      responses:
        '200':
          description: Team members
    post:
      tags: [Team Members]
      summary: Create team member and user account
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/TeamMemberCreateRequest'
      responses:
        '201':
          description: Team member created
  /team-members/{id}:
    get:
      tags: [Team Members]
      summary: Get team member by id
      parameters:
        - $ref: '#/components/parameters/IdParam'
      responses:
        '200':
          description: Team member
    put:
      tags: [Team Members]
      summary: Update team member and linked user
      parameters:
        - $ref: '#/components/parameters/IdParam'
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/TeamMemberUpdateRequest'
      responses:
        '200':
          description: Team member updated
    delete:
      tags: [Team Members]
      summary: Delete team member and linked user
      parameters:
        - $ref: '#/components/parameters/IdParam'
      responses:
        '200':
          description: Team member deleted
...
components:
  parameters:
    IdParam:
      in: path
      name: id
      required: true
      schema:
        type: integer
        format: int64
  schemas:
...
    TeamMemberCreateRequest:
      type: object
      properties:
        teamMemberName: { type: string }
        teamMemberMobile: { type: string }
        teamMemberEmail: { type: string }
        teamMemberRole: { type: integer, format: int64 }
        password: { type: string }
    TeamMemberUpdateRequest:
      type: object
      properties:
        teamMemberName: { type: string }
        teamMemberMobile: { type: string }
        teamMemberEmail: { type: string }
        teamMemberRole: { type: integer, format: int64 }